change JOBS value to 1
Since each instance of `cc1plus` was using close to 1.7G in tests,
a value of greater than one will likely cause swapping. On a micro-sd
this can be bad for the system. Setting jobs to 1 may build slower
overall, but going from ddr4 to microsd swapfile is like hitting a brick
wall.